댓글 삭제 추가, 글 삭제 post 방식으로 변경

+++ myApp/views.py
@@ -25,20 +25,48 @@ def detail(request, post_id):
     post = get_object_or_404(Post, id=post_id)
     return render(request, 'detail.html', {'post': post})
     
+# myApp/views.py
+
+from django.shortcuts import render, redirect, get_object_or_404, reverse
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.models import User
+from django.utils import timezone
+from .models import Post, Comment
+
+def index(request):
+    posts = Post.objects.order_by('-id')
+    app_url = request.path
+    return render(request, 'index.html', {'posts':posts, 'app_url':app_url})
+
+def post(request):
+    if request.method == 'POST':
+        post = Post()
+        post.main_text = request.POST['main_text']
+        post.create_user = User.objects.get(username = request.user.get_username())
+        post.create_date = timezone.datetime.now()
+        post.save()
+        return redirect(reverse('index'))
+    return render(request, 'post.html')
+
+def detail(request, post_id):
+    post = get_object_or_404(Post, id=post_id)
+    return render(request, 'detail.html', {'post': post})
+    
 def update(request, post_id):
     post = Post.objects.get(id = post_id)
     if request.method == 'POST':
         post.main_text = request.POST['main_text']
         post.create_user = User.objects.get(username = request.user.get_username())
-        post.update_date = timezone.datetime.now()
         post.save()
         return redirect(reverse('index'))
     return render(request, 'update.html')
 
 def delete(request, post_id):
-    post = get_object_or_404(Post, id=post_id)
-    post.delete()
-    return redirect(reverse('index'))
+    post = Post.objects.get(id = post_id)
+    if request.method == 'POST':
+        post = get_object_or_404(Post, id=post_id)
+        post.delete()
+        return redirect(reverse('index'))
 
 @login_required
 def c_post(request, post_id):
@@ -48,6 +76,47 @@ def c_post(request, post_id):
         comment_user = User.objects.get(username = request.user.get_username())
         Comment.objects.create(comment=comment, comment_text=comment_text, comment_user=comment_user)
 
+        if request.POST.get('app_url') == '/myApp/index/':
+            return redirect(reverse('index'), post_id)
+        else :
+            return redirect('detail', post_id)
+
+@login_required
+def c_delete(request, post_id, comment_id):
+    post = get_object_or_404(Post, id=post_id)
+    comment = get_object_or_404(Comment, id=comment_id)
+
+    if request.method =='POST':
+            
+        comment.delete()
+        
+        if request.POST.get('app_url') == '/myApp/index/':
+            return redirect(reverse('index'), post_id)
+        else :
+            return redirect('detail', post_id)
+
+@login_required
+def c_post(request, post_id):
+    if request.method =='POST':
+        comment = get_object_or_404(Post, id=post_id)
+        comment_text = request.POST.get('comment_text')
+        comment_user = User.objects.get(username = request.user.get_username())
+        Comment.objects.create(comment=comment, comment_text=comment_text, comment_user=comment_user)
+
+        if request.POST.get('app_url') == '/myApp/index/':
+            return redirect(reverse('index'), post_id)
+        else :
+            return redirect('detail', post_id)
+
+@login_required
+def c_delete(request, post_id, comment_id):
+    post = get_object_or_404(Post, id=post_id)
+    comment = get_object_or_404(Comment, id=comment_id)
+
+    if request.method =='POST':
+            
+        comment.delete()
+        
         if request.POST.get('app_url') == '/myApp/index/':
             return redirect(reverse('index'), post_id)
         else :
